Move KeystoneSDKConfig from sdk to types for modular import
