+++ src/types/config.ts
@@ -1,0 +1,4 @@
+
+export interface KeystoneSDKConfig {
+    origin?: string
+  }
