+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from './props'
 export * from './signature'
 export * from './token'
 export * from './ur'
+export * from './config'
